+++ src/lib.rs
@@ -1,6 +1,6 @@
 //! # Ferinth
 //!
-//! Ferinth is a simple library for using the [Modrinth API](https://github.com/modrinth/labrinth/wiki/API-Documentation) in Rust projects.
+//! Ferinth is a simple library for using the [Modrinth API](https://github.com/modrinth/labrinth/wiki/API-Documentation) in Rust.
 //! It uses [Reqwest](https://docs.rs/reqwest/) as its HTTPS client and deserialises responses to strongly typed structs using [SerDe](https://serde.rs/).
 //!
 //! ## Features
@@ -8,7 +8,7 @@
 //! This crate includes the following:
 //!
 //! - All structure definitions based on <https://docs.modrinth.com/api-spec/>
-//! - All of the GET calls
+//! - All of the GET and POST calls that don't require authentication
 //!
 //! This crate uses [RusTLS](https://docs.rs/rustls/) rather than OpenSSL, because OpenSSL is outdated and slower.
 //!
@@ -21,6 +21,8 @@ mod api_calls;
 mod request;
 pub mod structures;
 mod url_join_ext;
+
+use reqwest::Client;
 
 #[derive(thiserror::Error, Debug)]
 pub enum Error {
@@ -50,7 +52,40 @@ pub(crate) type Result<T> = std::result::Result<T, Error>;
 /// let sodium_mod = modrinth.get_project("sodium").await?;
 /// # Ok(()) }
 /// ```
-#[derive(Debug, Clone, Default)]
+#[derive(Debug, Clone)]
 pub struct Ferinth {
-    client: reqwest::Client,
+    client: Client,
+}
+
+impl Default for Ferinth {
+    fn default() -> Self {
+        Self {
+            client: Client::builder()
+                .user_agent(concat!(
+                    env!("CARGO_CRATE_NAME"),
+                    "/",
+                    env!("CARGO_PKG_VERSION")
+                ))
+                .build()
+                .unwrap(),
+        }
+    }
+}
+
+impl Ferinth {
+    /// Instantiate the container with the provided [user agent](https://docs.modrinth.com/api-spec/#section/User-Agents).
+    /// `program_name` is required, and the `version` and `contact` are optional but recommended by Modrinth.
+    pub fn new(program_name: &str, version: Option<&str>, contact: Option<&str>) -> Self {
+        Self {
+            client: Client::builder()
+                .user_agent(format!(
+                    "{}{}{}",
+                    program_name,
+                    version.map_or("".into(), |version| format!("/{}", version)),
+                    contact.map_or("".into(), |contact| format!(" ({})", contact))
+                ))
+                .build()
+                .unwrap(),
+        }
+    }
 }